requirements.txt
Flask==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
python-dotenv==0.10.3
psycopg2==2.8.4
